video: add a feature to upgrade the video board firmware via Recovery
Fix a bug in the keyboard driver that resulted in corrupted NOR flash.
On boot, if the video board is in recovery mode, then the bootloader will
automatically run the flasher on the UART.
Add a new entry in the menu to flash the video board firmware. The firmware needs
to be sent via UART, just like NOR flash programming.