video: add support for the Zeal 8-bit Video Card
It is now possible to use the bootloader with the Zeal 8-bit video card and the PS/2 keyboard.
The standard output can now be chosen in a configuration file.
Note that the UART is still enabled even if it not the standard output because
it is required to receive files.